requirements.txt
pytube3==9.6.4
selenium==3.141.0
bs4==0.0.1
requests==2.23.0
Flask==1.1.2
mongoengine==0.20.0